+++ PaulEngine/src/Platform/OpenGL/OpenGLRenderAPI.h
@@ -8,7 +8,7 @@ namespace PaulEngine {
 		~OpenGLRenderAPI();
 
 		virtual void Init() override;
-		virtual void SetViewport(const glm::vec2& position, const glm::vec2& size) override;
+		virtual void SetViewport(const glm::ivec2& position, const glm::ivec2& size) override;
 		virtual void SetClearColour(const glm::vec4& colour) override;
 		virtual void Clear() override;
